+++ Domain/Entities/HistorialMedico.cs
@@ -5,11 +5,11 @@ namespace Domain.Entities;
 
 public partial class HistorialMedico
 {
-    public int IdHistorial { get; set; }
+    public decimal IdHistorial { get; set; }
 
-    public long? IdPaciente { get; set; }
+    public decimal? IdPaciente { get; set; }
 
-    public long? IdMedico { get; set; }
+    public decimal? IdMedico { get; set; }
 
     public DateTime Fecha { get; set; }
